+++ mosakin-frontend/src/components/organisms/paidListRows/index.stories.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { PaidListRows } from ".";
+import { storiesOf } from "@storybook/react";
+import { rows } from "@/models/models/paidList/mock";
+
+storiesOf("organisms/PaidListRows", module).add("PaidListRows", () => {
+  return <PaidListRows rows={rows} />;
+});
